Troca geração de propriedades para server side
