+++ client/.env.example
@@ -1,4 +1,4 @@
-SITE_URL=http://localhost:3000
+VERCEL_URL=http://localhost:3000
 FIREBASE_PROJECT_ID=
 FIREBASE_PRIVATE_KEY=
 FIREBASE_CLIENT_EMAIL=
