+++ client/pages/_app.js
@@ -1,6 +1,4 @@
-import App from 'next/app'
 import Head from 'next/head'
-import Nav from '../components/Nav'
 
 import '../node_modules/bulma/bulma.sass'
 import '../styles/globals.css'
@@ -12,24 +10,9 @@ const Caju = ({ Component, pageProps, pages }) => {
         <title>Caju Cozinha Independente</title>
         <link rel="icon" href="/images/logo-caju-cozinha-independente-porto-alegre.png" />
       </Head>
-      <Nav pages={pages} />
       <Component {...pageProps} />
     </>
   )
-}
-
-Caju.getInitialProps = async (appContext) => {
-  const appProps = await App.getInitialProps(appContext);
-
-  const res = await fetch(`${process.env.VERCEL_URL}/api/paginas`);
-  const json = await res.json();
-
-  const pages = {
-    produtos: json.filter(({ secao }) => secao === 'produtos'),
-    comoFunciona: json.filter(({ secao }) => secao === 'como-funciona'),
-  }
-
-  return { ...appProps, pages }
 }
 
 export default Caju
